Refactor(web-react): Clean up shared non-dictionary types and constants
