+++ packages/web-react/src/constants/position.ts
@@ -1,4 +1,4 @@
-export const Alignment = {
+export const Position = {
   LEFT: 'left',
   RIGHT: 'right',
   TOP: 'top',
