--- packages/web-react/src/types/shared/alignments.ts
@@ -1,4 +1,0 @@
-import { Alignment } from '../../constants';
-
-export type AlignmentDictionaryKeys = keyof typeof Alignment;
-export type AlignmentDictionaryType = (typeof Alignment)[AlignmentDictionaryKeys];
