--- packages/web-react/src/types/shared/colors.ts
@@ -1,6 +1,0 @@
-export type Color = 'primary' | 'secondary' | 'tertiary' | 'inverted' | 'success' | 'danger';
-
-export interface ColorProps {
-  /** The color of the component. */
-  color?: Color;
-}
