+++ packages/web-react/src/types/shared/positions.ts
@@ -1,0 +1,4 @@
+import { Position } from '../../constants';
+
+export type PositionKeys = keyof typeof Position;
+export type PositionType = (typeof Position)[PositionKeys];
